# openapi.yaml
openapi: 3.0.3
info:
  title: SOC Application API
  description: |
    Security Operations Center (SOC) application for analyzing Zscaler NSS web proxy logs.
    
    This API provides endpoints for:
    - User authentication
    - Log file upload and management
    - Log entry analysis and filtering
    - Dashboard statistics and analytics
    - Anomaly detection
    - AI-powered threat analysis
    
    ## Authentication
    Most endpoints require JWT authentication. Obtain a token by calling the `/api/auth/login` endpoint,
    then include it in the `Authorization` header as `Bearer <token>`.
    
  version: 1.0.0
  contact:
    name: SOC Application Support
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /api/dashboard/stats:
    get:
      tags:
        - Dashboard
      summary: Get dashboard statistics
      description: Get overall statistics for the dashboard
      operationId: getDashboardStats
      parameters:
        - name: log_file_id
          in: query
          description: Filter statistics by log file ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/dashboard/timeline:
    get:
      tags:
        - Dashboard
      summary: Get timeline data
      description: Get timeline data for charts showing activity over time
      operationId: getTimeline
      parameters:
        - name: log_file_id
          in: query
          description: Filter by log file ID
          schema:
            type: string
            format: uuid
        - name: bucket_minutes
          in: query
          description: Time bucket size in minutes
          schema:
            type: integer
            default: 15
            minimum: 1
        - name: hours
          in: query
          description: Number of hours to include
          schema:
            type: integer
            default: 24
            minimum: 1
      responses:
        '200':
          description: Timeline data
          content:
            application/json:
              schema:
                type: object
                properties:
                  buckets:
                    type: array
                    items:
                      $ref: '#/components/schemas/TimelineBucket'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/dashboard/top-categories:
    get:
      tags:
        - Dashboard
      summary: Get top URL categories
      description: Get the most frequently accessed URL categories
      operationId: getTopCategories
      parameters:
        - name: log_file_id
          in: query
          description: Filter by log file ID
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Number of top categories to return
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Top URL categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/CategoryStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/dashboard/top-domains:
    get:
      tags:
        - Dashboard
      summary: Get top domains
      description: Get the most frequently accessed domains
      operationId: getTopDomains
      parameters:
        - name: log_file_id
          in: query
          description: Filter by log file ID
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Number of top domains to return
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Top domains
          content:
            application/json:
              schema:
                type: object
                properties:
                  domains:
                    type: array
                    items:
                      $ref: '#/components/schemas/DomainStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/dashboard/top-users:
    get:
      tags:
        - Dashboard
      summary: Get top users
      description: Get users with the highest request counts
      operationId: getTopUsers
      parameters:
        - name: log_file_id
          in: query
          description: Filter by log file ID
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Number of top users to return
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Top users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/dashboard/recent-logs:
    get:
      tags:
        - Dashboard
      summary: Get recent log entries
      description: Get the most recent log entries
      operationId: getRecentLogs
      parameters:
        - name: log_file_id
          in: query
          description: Filter by log file ID
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Number of recent logs to return
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Recent log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/LogEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
...
  schemas:
...
    LogEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        log_file_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        location:
          type: string
          example: ny-gre
        protocol:
          type: string
          example: HTTPS
        url:
          type: string
          example: https://example.com/path
        domain:
          type: string
          example: example.com
        action:
          type: string
          enum: [Allowed, Blocked]
        app_name:
          type: string
          example: Gmail
        app_class:
          type: string
          example: Business Apps
        throttle_req_size:
          type: integer
        throttle_resp_size:
          type: integer
        req_size:
          type: integer
        resp_size:
          type: integer
        url_class:
          type: string
          example: Business Use
        url_supercat:
          type: string
          example: Business and Economy
        url_cat:
          type: string
          example: Email
        dlp_dict:
          type: string
        dlp_eng:
          type: string
        dlp_hits:
          type: integer
        file_class:
          type: string
        file_type:
          type: string
        location2:
          type: string
        department:
          type: string
          example: Engineering
        client_ip:
          type: string
          format: ipv4
          example: 172.17.3.49
        server_ip:
          type: string
          format: ipv4
          example: 66.211.175.229
        http_method:
          type: string
          example: GET
        http_status:
          type: integer
          example: 200
        user_agent:
          type: string
          example: Mozilla/5.0
        threat_category:
          type: string
          example: Phishing
        fw_filter:
          type: string
        fw_rule:
          type: string
        policy_type:
          type: string
        reason:
          type: string
        is_anomalous:
          type: boolean
        anomaly_type:
          type: string
          enum: [burst_blocked, malicious_domain, risky_category, unusual_ua, large_download]
        anomaly_reason:
          type: string
        anomaly_confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
        created_at:
          type: string
          format: date-time
    DashboardStats:
      type: object
      properties:
        total_requests:
          type: integer
          example: 100000
        blocked_events:
          type: integer
          example: 1250
        malicious_urls:
          type: integer
          example: 45
        high_risk_users:
          type: integer
          example: 3
        data_transfer_bytes:
          type: integer
          format: int64
          example: 5368709120
        trends:
          type: object
          properties:
            total_requests_pct:
              type: number
              format: float
              example: 5.2
            blocked_events_pct:
              type: number
              format: float
              example: -2.1
    TimelineBucket:
      type: object
      properties:
        time:
          type: string
          format: date-time
          example: "2022-06-20T12:00:00Z"
        total:
          type: integer
          example: 1500
        blocked:
          type: integer
          example: 25
    CategoryStats:
      type: object
      properties:
        name:
          type: string
          example: Email
        count:
          type: integer
          example: 15000
        percentage:
          type: number
          format: float
          example: 15.5
    DomainStats:
      type: object
      properties:
        domain:
          type: string
          example: example.com
        count:
          type: integer
          example: 5000
        blocked_count:
          type: integer
          example: 10
    UserStats:
      type: object
      properties:
        identifier:
          type: string
          example: Engineering
        request_count:
          type: integer
          example: 25000
        risk_score:
          type: number
          format: float
          example: 35.5
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...